+++ app/src/main/res/layout/activity_detail_laporan.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/bggradient"
     android:orientation="vertical"
-    tools:context=".detaillaporan.DetailLaporanActivity">
+    tools:context=".view.detaillaporan.DetailLaporanActivity">
 
     <LinearLayout
         android:paddingVertical="20dp"
