+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bggradient"
-    tools:context=".LoginActivity">
+    tools:context=".view.auth.LoginActivity">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/batas_atas"
