+++ app/src/main/res/layout/activity_menu_rt.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bggradient"
-    tools:context=".HomeActivity">
+    tools:context=".view.home.HomeActivity">
 
     <FrameLayout
         android:id="@+id/container"
