+++ app/src/main/res/layout/fragment_riwayat.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".menu.UlasanAdapter">
+    tools:context=".view.menu.UlasanAdapter">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
