+++ app/src/main/res/layout/fragment_ulasan_rt.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".menurt.ulasan.UlasanRTFragment">
+    tools:context=".view.menu.rt.ulasan.UlasanRTFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
